Move scan actions to app context and interactor
